streamlit_app/requirements.txt
streamlit==1.34.0
extra-streamlit-components==0.1.60
pandas===2.0.3
matplotlib==3.7.2
ipython==8.14.0
numpy==1.24.3
seaborn==0.12.2
nltk==3.8.1
scikit-learn==1.3.0
Pillow==10.0.0
filesplit==4.0.1
regex==2023.6.3
pydot==2.0.0
https://files.pythonhosted.org/packages/cc/58/96aff0e5cb8b59c06232ea7e249ed902d04ec89f52636f5be06ceb0855fe/extra_streamlit_components-0.1.60-py3-none-any.whl
streamlit-option-menu==0.3.6
fastapi==0.103.0
uvicorn==0.20.0
chardet==5.2.0
filesplit==4.0.1
pydot==2.0.0
passlib==1.7.4
pyjwt==2.4.0
python-multipart==0.0.9
mysql-connector-python==8.3.0
pytest==7.4.0
python-dotenv==1.0.0
altair[all]==5.3.0
pillow==10.0.0
